Code review integration for 86379933aa486c9e81cef2ab3113a68d2dced360
Throw an exception in updateBCD if the BCD is already set
Simplify BCD update (Long v.s int) for consistency)

+++ account/src/main/java/org/killbill/billing/account/dao/AccountDao.java
@@ -51,5 +51,5 @@ public interface AccountDao extends EntityDao<AccountModelDao, Account, AccountA
 
     List<AccountEmailModelDao> getEmailsByAccountId(UUID accountId, InternalTenantContext context);
 
-    Long getAccountBCD(UUID accountId, InternalTenantContext context);
+    Integer getAccountBCD(UUID accountId, InternalTenantContext context);
 }
